adhoc fix session leak
