+++ query/src/sessions/query_ctx_shared.rs
@@ -19,13 +19,11 @@ use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
 
-use chrono_tz::Tz;
 use common_base::base::Progress;
 use common_base::base::Runtime;
 use common_contexts::DalContext;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_io::prelude::FormatSettings;
 use common_meta_types::UserInfo;
 use common_planners::PlanNode;
 use common_users::RoleCacheMgr;
@@ -40,7 +38,6 @@ use crate::catalogs::CatalogManager;
 use crate::clusters::Cluster;
 use crate::servers::http::v1::HttpQueryHandle;
 use crate::sessions::Session;
-use crate::sessions::SessionType;
 use crate::sessions::Settings;
 use crate::sql::SQLCommon;
 use crate::storages::Table;
@@ -288,34 +285,6 @@ impl QueryContextShared {
 
     pub fn get_config(&self) -> Config {
         self.session.get_config()
-    }
-
-    pub fn get_format_settings(&self) -> Result<FormatSettings> {
-        let settings = self.get_settings();
-        let mut format = FormatSettings::default();
-        if let SessionType::HTTPQuery = self.session.get_type() {
-            format.false_bytes = vec![b'f', b'a', b'l', b's', b'e'];
-            format.true_bytes = vec![b't', b'r', b'u', b'e'];
-        }
-        {
-            format.record_delimiter = settings.get_record_delimiter()?;
-            format.field_delimiter = settings.get_field_delimiter()?;
-            format.empty_as_default = settings.get_empty_as_default()? > 0;
-            format.skip_header = settings.get_skip_header()? > 0;
-
-            let tz = String::from_utf8(settings.get_timezone()?).map_err(|_| {
-                ErrorCode::LogicalError("Timezone has been checked and should be valid.")
-            })?;
-            format.timezone = tz.parse::<Tz>().map_err(|_| {
-                ErrorCode::InvalidTimezone("Timezone has been checked and should be valid")
-            })?;
-
-            let compress = String::from_utf8(settings.get_compression()?).map_err(|_| {
-                ErrorCode::UnknownCompressionType("Compress type must be valid utf-8")
-            })?;
-            format.compression = compress.parse()?
-        }
-        Ok(format)
     }
 
     pub fn get_connection_id(&self) -> String {
